myPage - 별점 표시 수정, ReviewItemDTO ratingScore 자료형 변경, 400.html 추가

+++ src/main/java/project/gasnow/user/model/dto/ReviewItemDTO.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class ReviewItemDTO {
     private String createdAt;
     private String stationName;
-    private int ratingScore;
+    private Double ratingScore;
 }
